backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = [".."]
include = ["backend*"]

[project]
name = "greenroom-backend"
version = "0.1.0"
description = "DC Concert Aggregator — Flask REST API"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0,<4",
    "sqlalchemy>=2.0,<3",
    "alembic>=1.13,<2",
    "psycopg2-binary>=2.9,<3",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "celery[redis]>=5.3,<6",
    "redis>=5.0,<6",
    "requests>=2.31,<3",
    "beautifulsoup4>=4.12,<5",
    "lxml>=5.0,<6",
    "PyJWT[crypto]>=2.8,<3",
    "gunicorn>=21.2,<23",
    "webauthn>=2.0,<3",
    "knuckles-client>=0.1.0,<0.2.0",
    "sentry-sdk[flask,celery,sqlalchemy]>=2.0,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-cov>=5.0,<6",
    "black>=24.0,<25",
    "ruff==0.15.11",
    "mypy>=1.9,<2",
    "types-requests>=2.31,<3",
    "types-beautifulsoup4>=4.12,<5",
    "responses>=0.25,<1",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["backend"]

[tool.ruff.lint.per-file-ignores]
# Test files commonly mix runtime and annotation use of pytest/typing imports;
# TYPE_CHECKING blocks become brittle (add a @pytest.fixture later and the
# test breaks at import time). The minor runtime-import cost is irrelevant
# in a test suite.
"tests/**/*.py" = ["TCH001", "TCH002", "TCH003"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
# Tests monkey-patch module-level imports (e.g. `module.repo_alias`),
# which strict's default `no_implicit_reexport` would flag as attr-defined
# on every patch target. The alternative — adding `__all__` everywhere —
# is noise for internal modules.
implicit_reexport = true
exclude = ["migrations/"]
disable_error_code = [
    "comparison-overlap",
    "list-item",
    "union-attr",
    "unused-ignore",
    # The error codes below fire almost exclusively in test modules that
    # use small `_Fake*` dataclasses as stand-ins for SQLAlchemy models.
    # Strict checks around identity and container element types can't see
    # through the monkey-patches the tests use; these codes are low-value
    # in source where we catch the same bugs via stricter narrowing.
    "arg-type",
    "operator",
    "call-overload",
    "func-returns-value",
    "import-untyped",
]

[[tool.mypy.overrides]]
module = ["celery.*"]
ignore_missing_imports = true
follow_imports = "skip"

# Tests use small `_Fake*` dataclasses as stand-ins for SQLAlchemy models
# (the repos are monkey-patched to return them). Strict checks around
# identity and container element types can't see through the monkey-patch,
# so we relax `arg-type`, `operator`, `call-overload`, and
# `func-returns-value` for test modules only — production code keeps the
# strict checks.

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--strict-markers -v"

[tool.coverage.run]
source = ["backend"]
omit = [
    "*/tests/*",
    "*/scripts/*",
    "*/migrations/*",
    "*/celery_app.py",
    "*/app.py",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
